backend-python/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pinecone
groq
pydantic
lxml
curl_cffi
numpy
rank_bm25
huggingface_hub
